PaymentProviderで行っていた処理をHomeコンポーネントに移動
